Collision_handler har fått en uppgradering
objekt, runda och rektangulära kan nu fråga alla objekt i cover_listan
om kollision sker, samt med fönster storlekarna

+++ gameobject.h
@@ -11,8 +11,8 @@ public:
     virtual void rotate(int turn_direction); //turn_direction är 0 eller +-1 för att avgöra om
                                              //objektet ska vridas och åt vilket håll.
     virtual std::string get_name(){return name;}
-    virtual int get_xpos(){return xpos;}
-    virtual int get_ypos(){return ypos;}
+    virtual double get_xpos(){return xpos;}
+    virtual double get_ypos(){return ypos;}
     virtual double get_direction(){return direction;}
 
 protected:
